Move syntax tree classes to interpreter module
